% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper_fpds_data.R
\name{clean_text}
\alias{clean_text}
\title{Clean Text}
\usage{
clean_text(text)
}
\arguments{
\item{text}{The text to be cleaned.}
}
\value{
Cleaned text.
}
\description{
This function cleans the text.
}
