% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper_fpds_data.R
\name{scrape_data_from_all_pages}
\alias{scrape_data_from_all_pages}
\title{Scrape Data from All Pages}
\usage{
scrape_data_from_all_pages(url, search_term)
}
\arguments{
\item{url}{The URL for the page.}

\item{search_term}{The search term.}
}
\value{
A data frame containing the scraped data from all pages.
}
\description{
This function scrapes data from all pages.
}
