% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper_fpds_data.R
\name{scrape_fpds_data}
\alias{scrape_fpds_data}
\title{Scrape FPDS Data from URL}
\usage{
scrape_fpds_data(url)
}
\arguments{
\item{url}{The URL from which to scrape data.}
}
\value{
A data frame containing the scraped data.
}
\description{
This function scrapes data from the FPDS website based on a provided URL.
}
