% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper_fpds_data.R
\name{url_exists}
\alias{url_exists}
\title{Check if URL exists}
\usage{
url_exists(url)
}
\arguments{
\item{url}{The URL to check.}
}
\value{
TRUE if the URL exists, FALSE otherwise.
}
\description{
This function checks if the given URL exists.
}
