% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper_fpds_data.R
\name{verify_fpds_url}
\alias{verify_fpds_url}
\title{Verify FPDS URL}
\usage{
verify_fpds_url(url)
}
\arguments{
\item{url}{The URL to verify.}
}
\value{
TRUE if the URL is standard according to FPDS, FALSE otherwise.
}
\description{
This function verifies if the given URL is a standard FPDS URL.
}
